Move GamePiece from gameplay elements to a separated package.
| + Added package `gamepiece`.
|_______________________________________________________________________
