+++ src/gameplay/Level.java
@@ -1,0 +1,5 @@
+package gameplay;
+
+public class Level {
+
+}
